refactor(terrarium): split remote recipe operations
Move whole-recipe RPC handling into a focused service mixin so the remote facade stays below the repository hard file-size limit without changing its public API.
